star coding TV Details page

+++ Models/TmDb/MovieDetail.cs
@@ -87,7 +87,7 @@ namespace ReelJunkies.Models.TmDb
         public string name { get; set; }
         public string username { get; set; }
         public string avatar_path { get; set; }
-        public int? rating { get; set; }
+        public decimal? rating { get; set; }
     }
